refactoring for the new services - added tests too

+++ public/js/application/services/BookService.js
@@ -1,6 +1,6 @@
 "use strict";
 
-myClass.factory('BookService', ['$http', function($http)
+myClass.factory('BookService', ['$http', 'lib', function($http, lib)
 {
     var URL = '/api/books';
 
@@ -11,16 +11,28 @@ myClass.factory('BookService', ['$http', function($http)
 
     function _registerBook(livro)
     {
+        if (lib.isObjectInvalid(livro))
+            throw new Error('Não é possível cadastrar o livro. Parâmetros informados de forma errada.');
+
         return $http.post(URL, livro);
     }
 
     function _editBook(id, livro)
     {
+        if (lib.isStringInvalid(id))
+            throw new Error('Não é possível editar o livro. Id informado de forma errada.');
+
+        if (lib.isObjectInvalid(livro))
+            throw new Error('Não é possível editar o livro. Livro informado de forma errada.');
+
         return $http.put(URL + '/' + id, livro);
     }
 
     function _deleteBook(id)
     {
+        if (lib.isStringInvalid(id))
+            throw new Error('Não é possível deletar o livro. Id informado de forma errada.');
+
         return $http.delete(URL + '/' + id);
     }
