+++ public/js/application/services/ClazzService.js
@@ -1,6 +1,6 @@
 "use strict";
 
-myClass.factory('ClazzService', ['$http', function($http)
+myClass.factory('ClazzService', ['$http', 'lib', function($http, lib)
 {
     var URL = '/api/classes';
 
@@ -16,16 +16,28 @@ myClass.factory('ClazzService', ['$http', function($http)
 
     function _registerClazz(turma)
     {
+        if (lib.isObjectInvalid(turma))
+            throw new Error('Não é possível cadastrar a turma. O paramêtro foi passado de forma errada.');
+
         return $http.post(URL, turma);
     }
 
     function _editClazz(id, turma)
     {
+        if (lib.isObjectInvalid(turma))
+            throw new Error('Não é possível editar a turma. O paramêtro TURMA foi passado de forma errada.');
+
+        if (lib.isStringInvalid(id))
+            throw new Error('Não é possível editar a turma. O paramêtro ID foi passado de forma errada.');
+
         return $http.put(URL + '/' + id, turma);
     }
 
     function _deleteClazz(id)
     {
+        if (lib.isStringInvalid(id))
+            throw new Error('Não é possível deletar a turma. O paramêtro ID foi passado de forma errada.');
+
         return $http.delete(URL + '/' + id);
     }
